+++ apps/docs/vite.config.ts
@@ -1,0 +1,19 @@
+import vue from "@vitejs/plugin-vue";
+import { fileURLToPath, URL } from "node:url";
+import { defineConfig } from "vite";
+
+// https://vitejs.dev/config
+export default defineConfig({
+  plugins: [vue()],
+  resolve: {
+    alias: {
+      "~components": getFilePath("../../packages/sit-onyx/src/components"),
+      "@": getFilePath("../../packages/sit-onyx/src"),
+    },
+  },
+});
+
+/** Gets the given path while ensuring cross-platform and correct decoding */
+function getFilePath(path: string) {
+  return fileURLToPath(new URL(path, import.meta.url));
+}